requirements.txt
python-telegram-bot
flask
redis_wrap
requests
rq
certifi